E - Updated clippy cleanup

+++ src/db/db_persister.rs
@@ -115,7 +115,7 @@ impl Persister for DatabasePersister {
                 .values(cols)
                 .and_where(
                     Expr::col(Alias::new(DEFAULT_ID_COLUMN_NAME).into_iden())
-                        .eq::<u64>(id.unwrap() as u64),
+                        .eq::<u64>(id.unwrap()),
                 )
                 .to_owned();
 
@@ -128,7 +128,7 @@ impl Persister for DatabasePersister {
                 .values(vals)?
                 .to_owned();
             let result = db.execute(builder.build(&stmt)).await?;
-            Ok(result.last_insert_id() as u64)
+            Ok(result.last_insert_id())
         }
     }
 }
